feat(ui): Show file size with units (#41)
* refactor: Create barrel file in lib folder

* feat(ui): Show size with units
